refactor: AudioVolumeAnalyzerクラスを分離
